Rename
The repository is hereby renamed with its internal name `Lothlorien` due to anonymous suggestion about a potential ambiguity with Minecraft `Far-Lands` mod. Other features are not altered so far. This project has nothing to do with any extent Minecraft mod or fan-fictions, as well as other IoT systems for mental healthcare.